fix(admin): move story structure QA lib into backend for Cloud Run
Docker image only ships backend/; lab service no longer imports scripts/
at module load time so preview and staging containers can start.
